feat: basic player on movie and serie
