+++ lib/features/serie/presentation/screens/serie_screen.dart
@@ -99,6 +99,7 @@ class _SerieScreenState extends ConsumerState<SerieScreen> {
                       pagingState: state,
                       fetchNextPage: fetchNextPage,
                       isListView: listViewMode,
+                      contentType: TrackContentType.serie,
                     ),
               ),
               onRefresh: () async {
